+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added more methods to the `ErcotAPI` class which uses the new [Ercot API](https://data.ercot.com/) for fetching data
   - Eventually, the `ErcotAPI` will be the primary way to fetch data from ERCOT, but for now, we still need the `Ercot` class because the new API doesn't support all datasets.
+- Add `pjm.get_gen_outages_by_type` to get generation outage data
 
 ## v0.28.0 - Mar 20, 2024